refactor(workspace): split lifecycle.rs and shared/afflictions.rs
