+++ game/src/tributes/lifecycle/mod.rs
@@ -1,0 +1,4 @@
+pub mod death;
+pub mod health;
+pub mod stamina;
+pub mod status;
