+++ shared/src/afflictions/source.rs
@@ -1,0 +1,60 @@
+use super::kind::{AfflictionKind, BodyPart};
+use serde::{Deserialize, Serialize};
+
+/// Storage discriminator. Same kind on different parts is independent;
+/// same kind on the same part collapses to one slot.
+pub type AfflictionKey = (AfflictionKind, Option<BodyPart>);
+
+/// Classification of trauma cause for mass casualty events.
+#[derive(Debug, Clone, Copy, PartialEq, Eq, Hash, Serialize, Deserialize)]
+#[serde(rename_all = "snake_case")]
+pub enum CauseClass {
+    Combat,
+    Environmental,
+    Mixed,
+}
+
+/// Specific cause of death, used in trauma source metadata.
+#[derive(Debug, Clone, PartialEq, Eq, Hash, Serialize, Deserialize)]
+#[serde(rename_all = "snake_case")]
+pub enum DeathCause {
+    Tribute(String),
+    Fire,
+    Drowning,
+    Starvation,
+    Dehydration,
+    Unknown,
+}
+
+/// Source of a trauma affliction, capturing the triggering event.
+#[derive(Debug, Clone, PartialEq, Eq, Hash, Serialize, Deserialize)]
+#[serde(rename_all = "snake_case")]
+pub enum TraumaSource {
+    WitnessedAllyDeath {
+        ally: String,
+        cause: Option<DeathCause>,
+    },
+    NearDeath {
+        cause: DeathCause,
+    },
+    Betrayal {
+        by: String,
+    },
+    MassCasualty {
+        cause_class: CauseClass,
+        deaths_this_cycle: u32,
+    },
+}
+
+/// Origin of an affliction. `Sponsor` and `Gamemaker` variants are reserved
+/// for future systems but ship in v1 to avoid enum churn (per spec §3).
+#[derive(Debug, Clone, PartialEq, Eq, Hash, Serialize, Deserialize)]
+#[serde(rename_all = "snake_case")]
+pub enum AfflictionSource {
+    Spawn,
+    Combat { attacker_id: String },
+    Environmental,
+    Cascade { from: AfflictionKey },
+    Sponsor,
+    Gamemaker,
+}
